rename module: kernel -> prover
